Added menu
Moved start game button to menu. When the menu is open, no other buttons can be clicked. At the end of the game, the menu displays the winner / loser and two buttons: 'Play Again' and 'Exit'. The 'Exit' button cannot actually exit the program at this time. It justs prints 'exit' to the terminal.